VNet Peering and Subnets in Azure
Azure VNet Peering provides the easy creation of multiple VNets to connect and communicate as one. Added to this, Subnets are used to segment resources within a VNet, providing a neat, secure, and scalable architecture. Low-latency, high-bandwidth communication can be enabled through VNet Peering hence, it is perfect for scenarios such as cross region data replication, scaling an application, and disaster recovery.